first draft after moving and subclassing item files
